avoid package missing error when doing pure pipeline validation

+++ common/clear_history.sh
@@ -55,7 +55,9 @@ fi
 # Remove Defender
 if [[ $distro == *"Ubuntu"* ]]
 then
-    apt-get purge -y mdatp
+    if dpkg -l | grep -qw mdatp; then
+        apt-get purge -y mdatp
+    fi
 else
     yum remove -y mdatp
 fi
